backend/requirements.txt
flask
flask-cors
pyjwt
bcrypt
python-dotenv
cryptography
